Upgrade Checkstyle to 10.2 and reconfigure checks
The latest version of Checkstyle is 10.2. At the same time, the
checkstyle.xml file is slightly disorganised.

Let's update the Checkstyle version in build.gradle, replace the
checkstyle.xml file with the one used in AddressBook 3 and add a
suppressions.xml file to support suppression of checks for test code
for developers intending to use JUnit.